Modification de la base de donnée pour prendre en compte les prises marquées comme prise (verbe prendre) par l'utilisateur et ajout de la fonctionnalité

+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/ListeTraitementsFragment.kt
@@ -167,7 +167,7 @@ class ListeTraitementsFragment : Fragment() {
                     for (prise in medoc.prises.split("/")) {
                         val traitementPrise: MutableList<String> = prise.split(";").toMutableList()
                         val maPrise = Prise(
-                            traitementPrise[0].toInt(),
+                            traitementPrise[0],
                             traitementPrise[1],
                             traitementPrise[2].toInt(),
                             traitementPrise[3]
